fix(engine): keep visual samples in unorm16
- Change VisualSample color payload from f32 RGBA to unorm16 RGBA
- Change visual sample points to integer texel coordinates for texture products
- Remove float normalization from fixture accumulation and solid texture test helpers

Follow-up: df8076a9

+++ lp-core/lpc-engine/src/visual_product/sample_request.rs
@@ -2,11 +2,11 @@
 
 use alloc::vec::Vec;
 
-/// Normalized or surface-space sample coordinate; exact interpretation is product-defined.
+/// Integer texel coordinate in the materialized visual surface.
 #[derive(Debug, Clone)]
 pub struct VisualSamplePoint {
-    pub x: f32,
-    pub y: f32,
+    pub x: u32,
+    pub y: u32,
 }
 
 #[derive(Debug, Clone)]
